src/requirements.txt
onnxruntime==1.16.3
numpy==1.26.4
Pillow==10.2.0
boto3==1.41.5 # local test için gerekli değil; lambda içinde olabilir
